Use oldest Channel out of the SimpleChannelPool on acquire
Motivation:

We previously used pollLast() to retrieve a Channel from the queue that backs SimpleChannelPool. This could lead to the problem that some Channels are very unfrequently used and so when these are used the connection was already be closed and so could not be reused.

Modifications:

Change pollLast() to poll() and so ensure Channels will be reused frequently.

Result:

Better Channel reuse.